Milvus-python/requirements.txt
pymilvus==2.5.0
python-dotenv>=1.0.1,<2.0.0
